Centralized mod integration init
